services/user/requirements.txt
-i https://pypi.org/simple

amqp==5.1.1

async-timeout==4.0.2

bcrypt==3.2.2

certifi==2022.6.15

cffi==1.15.0

charset-normalizer==2.0.12

click==8.1.3

colorama==0.4.5

deprecated==1.2.13

dnspython==2.2.1

eventlet==0.33.1

greenlet==2.0.0a2

idna==3.3

kombu==5.2.4

mock==4.0.3

nameko==3.0.0rc11

packaging==21.3

path==16.4.0

path.py==12.5.0

psycopg2-binary==2.9.3

pycparser==2.21

pyjwt==2.4.0

pyparsing==3.0.9

pyyaml==6.0

redis==4.3.3

requests==2.28.0

six==1.16.0

urllib3==1.26.9

vine==5.0.0

werkzeug==2.1.2

wrapt==1.14.1

